necessary changes by comments made Comment about LinkedList: Do you think that it's a good idea to add possibility to store null objects in the node? What if I will add null objects more then one time, how you will handle it and how you will implement indexOf? answer: Yeap, I think that it's good idea, cause, IMO, it'll be narrowing of List opportunities not to give user opportunity to add null elements. Due to my implementation, it'll return first null object index(and last, if it is lastIndefOf).
Comment about SetRunner: Is there any difference between upper and lower case?
answer: Nope, there is no difference, cause I think that it is not important if word is case-sensitive.

Comment about setRunner: Is it possible to have null element in the hash set?
answer: Yes, it's possible to have ONE null element in set.